+++ server/src/main/java/org/allaymc/server/world/AllayDimension.java
@@ -5,14 +5,17 @@ import lombok.extern.slf4j.Slf4j;
 import org.allaymc.api.block.dto.BlockStateWithPos;
 import org.allaymc.api.block.dto.PlayerInteractInfo;
 import org.allaymc.api.block.type.BlockState;
+import org.allaymc.api.debugshape.DebugShape;
 import org.allaymc.api.entity.Entity;
 import org.allaymc.api.entity.interfaces.EntityPlayer;
 import org.allaymc.api.eventbus.event.block.BlockBreakEvent;
 import org.allaymc.api.item.ItemStack;
 import org.allaymc.api.math.position.Position3i;
+import org.allaymc.api.network.ClientStatus;
 import org.allaymc.api.world.Dimension;
 import org.allaymc.api.world.DimensionInfo;
 import org.allaymc.api.world.generator.WorldGenerator;
+import org.allaymc.server.network.processor.impl.login.SetLocalPlayerAsInitializedPacketProcessor;
 import org.allaymc.server.world.service.AllayBlockUpdateService;
 import org.allaymc.server.world.service.AllayChunkService;
 import org.allaymc.server.world.service.AllayEntityService;
@@ -20,6 +23,7 @@ import org.allaymc.server.world.service.AllayLightService;
 import org.cloudburstmc.math.vector.Vector3f;
 import org.cloudburstmc.protocol.bedrock.data.LevelEvent;
 import org.cloudburstmc.protocol.bedrock.packet.LevelEventPacket;
+import org.cloudburstmc.protocol.bedrock.packet.ServerScriptDebugDrawerPacket;
 import org.jctools.maps.NonBlockingHashSet;
 import org.jetbrains.annotations.UnmodifiableView;
 
@@ -42,6 +46,7 @@ public class AllayDimension implements Dimension {
     protected final AllayBlockUpdateService blockUpdateService;
     protected final AllayLightService lightService;
     protected final Set<EntityPlayer> players;
+    protected final Set<DebugShape> debugShapes;
 
     public AllayDimension(AllayWorld world, WorldGenerator worldGenerator, DimensionInfo dimensionInfo) {
         this.world = world;
@@ -52,6 +57,7 @@ public class AllayDimension implements Dimension {
         this.blockUpdateService = new AllayBlockUpdateService(this);
         this.lightService = new AllayLightService(this);
         this.players = new NonBlockingHashSet<>();
+        this.debugShapes = new NonBlockingHashSet<>();
     }
 
     public void startTick() {
@@ -86,6 +92,27 @@ public class AllayDimension implements Dimension {
         this.players.add(player);
         this.chunkService.addChunkLoader(player);
         this.entityService.addEntity(player, runnable);
+        if (player.getClientStatus() == ClientStatus.IN_GAME) {
+            // Only send debug shapes to the players when they are in-game. This
+            // solves the issue that debug shapes won't be displayed if the player
+            // haven't fully joined. When the player join, the debug shapes will be
+            // sent to the player after the player is fully joined.
+            // See SetLocalPlayerAsInitializedPacketProcessor.
+            addDebugShapesTo(player);
+        }
+    }
+
+    /**
+     * Set this method to public because it is used in {@link SetLocalPlayerAsInitializedPacketProcessor}
+     */
+    public void addDebugShapesTo(EntityPlayer player) {
+        var packet = new ServerScriptDebugDrawerPacket();
+        for (var debugShape : debugShapes) {
+            // Let's send all the debug shapes in one packet to improve performance
+            debugShape.addViewer(player, false);
+            packet.getShapes().add(debugShape.toNetworkData());
+        }
+        player.sendPacket(packet);
     }
 
     @Override
@@ -102,12 +129,53 @@ public class AllayDimension implements Dimension {
             // Run the callback directly
             runnable.run();
         }
+        removeDebugShapesFrom(player);
+    }
+
+    protected void removeDebugShapesFrom(EntityPlayer player) {
+        var packet = new ServerScriptDebugDrawerPacket();
+        for (var debugShape : debugShapes) {
+            // Let's send all the remove notices in one packet to improve performance
+            debugShape.removeViewer(player, false);
+            packet.getShapes().add(debugShape.createRemovalNotice());
+        }
+        player.sendPacket(packet);
     }
 
     @Override
     @UnmodifiableView
     public Set<EntityPlayer> getPlayers() {
         return Collections.unmodifiableSet(this.players);
+    }
+
+    @Override
+    public void addDebugShape(DebugShape debugShape) {
+        this.debugShapes.add(debugShape);
+        for (var player : this.players) {
+            debugShape.addViewer(player);
+        }
+    }
+
+    @Override
+    public void removeDebugShape(DebugShape debugShape) {
+        this.debugShapes.remove(debugShape);
+        for (var player : this.players) {
+            debugShape.removeViewer(player);
+        }
+    }
+
+    @Override
+    @UnmodifiableView
+    public Set<DebugShape> getDebugShapes() {
+        return Collections.unmodifiableSet(this.debugShapes);
+    }
+
+    @Override
+    public void removeAllDebugShapes() {
+        for (var player : players) {
+            removeDebugShapesFrom(player);
+        }
+        this.debugShapes.clear();
     }
 
     @Override