+++ CHANGELOG.md
@@ -14,6 +14,11 @@ Unless otherwise specified, any version comparison below is the comparison of se
 
 <small>[Compare with 0.5.0](https://github.com/AllayMC/Allay/compare/0.5.0...HEAD)</small>
 
+### Added
+
+- (API) Added APIs for the new `DebugShape` system. Available debug shape classes can be found under `org.allaymc.api.debugshape`,
+  and add/remove methods for debug shape are available in `Dimension` class.
+
 ## 0.5.0 (API 0.8.0) - 2025-6-19
 
 <small>[Compare with 0.4.1](https://github.com/AllayMC/Allay/compare/0.4.1...HEAD)</small>